+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GitHub Pages documentation site (@markean, [#10](https://github.com/markean/aimz/issues/10)).
 - Installation instructions in the documentation site (@markean, [#10](https://github.com/markean/aimz/issues/10)).
 - `ArrayLoader` class supports `shuffle` and `drop_last` parameters for epoch training for `.fit()` (@markean, [#15](https://github.com/markean/aimz/issues/15)).
+- Added the `aimz.utils.data.collate` module with `jax_collate()`, which is now the default collate function for `ArrayLoader` (@markean, [#15](https://github.com/markean/aimz/issues/15)).
 
 ### Changed